Separating cusparse macros so the cusparse wrappers don't need to  be explicitly compiled w/ the handle
